Update test methods
