+++ FelicaLib_Remodeled/UnitTest45/FelicaTest.cs
@@ -2,10 +2,8 @@
 using Microsoft.VisualStudio.TestTools.UnitTesting;
 using System;
 using System.Linq;
-using System.Threading;
-using System.Threading.Tasks;
 
-namespace UnitTest45
+namespace UnitTest
 {
     [TestClass]
     public class FelicaTest
@@ -57,65 +55,6 @@ namespace UnitTest45
             {
                 var actual = felica.ReadWithoutEncryption(0x1317, 0);
                 Assert.AreEqual(12345, actual.ToEdyBalance());
-            }
-        }
-
-        [TestMethod]
-        public void ReadWithoutEncryption_ManyTimes1()
-        {
-            Task.Run(() =>
-            {
-                using (var felica = new Felica(FelicaSystemCode.Edy))
-                {
-                    for (int i = 0; i < 100; i++)
-                    {
-                        try
-                        {
-                            var data = felica.ReadWithoutEncryption(0x1317, 0);
-                            Console.WriteLine(data.ToEdyBalance());
-                        }
-                        catch (Exception ex)
-                        {
-                            Console.WriteLine(ex.Message);
-                        }
-                    }
-                }
-            });
-
-            Thread.Sleep(10000);
-        }
-
-        [TestMethod]
-        public void ReadWithoutEncryption_ManyTimes2()
-        {
-            Task.Run(() =>
-            {
-                for (int i = 0; i < 100; i++)
-                {
-                    try
-                    {
-                        Console.WriteLine("Start");
-                        ReadEdyBalanceEtc();
-                    }
-                    catch (Exception ex)
-                    {
-                        Console.WriteLine(ex.Message);
-                    }
-                }
-            });
-
-            Thread.Sleep(10000);
-        }
-
-        static void ReadEdyBalanceEtc()
-        {
-            using (var felica = new Felica(FelicaSystemCode.Edy))
-            {
-                Console.WriteLine(felica.GetIDm().ToHexString());
-                Console.WriteLine(felica.GetPMm().ToHexString());
-
-                var data = felica.ReadWithoutEncryption(0x1317, 0);
-                Console.WriteLine(data.ToEdyBalance());
             }
         }
     }
