+++ FelicaLib_Remodeled/UnitTest45/UnitTest45.csproj
@@ -57,6 +57,7 @@
     <Compile Include="Scenarios\NoPortTest.cs" />
     <Compile Include="Scenarios\EdyTest.cs" />
     <Compile Include="Scenarios\NoCardTest.cs" />
+    <Compile Include="FelicaLoadTest.cs" />
   </ItemGroup>
   <ItemGroup>
     <ProjectReference Include="..\FelicaLib_DotNet\FelicaLib_DotNet.csproj">
